Refactor to extract the score calculator
